+++ Chansooo/UIKitTutorial/preview.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
